move covariate balance calculation to another file
